% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{Centers variables for mixed effects models}
\usage{
center(
  dataset,
  x,
  grouping,
  type = "mean",
  standardize = FALSE,
  centerResult = FALSE
)
}
\arguments{
\item{dataset}{A dataset containing the variables to be centered and the grouping variable}

\item{x}{The variable or variables to be centered}

\item{grouping}{The variable or variables that define the grouping structure of the data}

\item{type}{a function to compute the grouping summary variable}

\item{standardize}{a logical value indicating whether x should be standardized before the computaion proceeds}

\item{centerResult}{a logical value indicating whether resulting grouping summary variable values should be centered at 0}
}
\value{
If \code{re_type} is "NA", the proportion of variance at the random effect is computed.

If re_type = "nested", the likeness of y scores in the same level 3 unit (the proportion of variance at Level3_factor), the likeness of y scores in the same level 2 units in the same level 3 unit (proportion of variance at Level3_factor and Level2_factor), and the likeness of level 2 units in the same level 3 unit (proportion of Level2_factor variance at Level3_factor) are computed.

If re_type = "cc", the likeness of y scores in the same C1_factor unit (correlation between outcome values of units in same C1_factor but different C2_factor), the likeness of y scores in the same C2_factor (correlation between outcome values of units in the same C2_factor but different C2_factor), and the likeness of y scores in the same C1_factor and C2_factor combination (correlation between outcome values of units in the same C2_factor and C2_factor) are computed.
}
\description{
Centers variables using the group-mean (person-mean) centering approach for mixed-effects models.
}
\references{
NEED
}
