% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimationsPlot.R
\name{estimationsPlot}
\alias{estimationsPlot}
\title{Plots estimations of the model}
\usage{
estimationsPlot(
  model,
  xlab = x,
  ylab = y,
  estimation_title = "Estimations Plot"
)
}
\arguments{
\item{model}{}

\item{xlab}{}

\item{ylab}{}

\item{estimation_title}{}
}
\description{
Plots the  estimations from the model
}
\examples{

\donttest{

}

}
\references{

}
