% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelCompare.R
\name{levelCompare}
\alias{levelCompare}
\title{Reports on a comparison of the mixed model vs. a model that does not account for nesting}
\usage{
levelCompare(model)
}
\arguments{
\item{model}{A linear mixed-effects model of class lmerMod or lmerModLmerTest}
}
\value{
Computes the AIC and BIC of the requested model and a model with the same predictors but absent the random intercept(s) and slope(s).
}
\description{
Uses AIC and BIC to compare whether the model that accounts for the correlation of responses within the same unit fits the data better than a model that assumes 0 correlation between responses within the same unit.
}
\examples{

\donttest{
data(instruction)
mod <- lme4::lmer(mathgain ~ (1 | classid), data = instruction)
levelCompare(mod)
}

}
\references{
Burnham, K. P., & Anderson, D. R. (2004). Multimodel inference: understanding AIC and BIC in model selection. Sociological methods & research, 33(2), 261-304.
Raftery, A. E. (1995). Bayesian model selection in social research. Sociological methodology, 111-163.
}
