% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelComparePlot.R
\name{levelComparePlot}
\alias{levelComparePlot}
\title{Plots comparison of accounting for nesting vs. not accounting for nesting}
\usage{
levelComparePlot(
  x,
  y,
  grouping,
  dataset,
  xlab = x,
  ylab = y,
  compare_title = "Comparison Plot",
  center = FALSE,
  paneled = FALSE
)
}
\arguments{
\item{x}{}

\item{y}{}

\item{grouping}{Grouping variable.}

\item{dataset}{A dataset containing the predictor, outcome, and grouping variables.}

\item{xlab}{Character vector specifying the horizontal axis label.}

\item{ylab}{Character vector specifying the vertical axis label.}

\item{compare_title}{Character vector specifying the title for the comparison plot.}

\item{center}{A logical value indicating whether x variable should be centered}

\item{paneled}{A logical value indicating whether the plot accounting for nesting should be split into panels}
}
\description{
Plots the line of best for the relationship between two variables accounting for nesting and not accounting for nesting.
}
\examples{

\donttest{

}

}
