% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm_assumptions.R
\name{mlm_assumptions}
\alias{mlm_assumptions}
\title{Reports the output of testing all assumptions for a multilevel model}
\usage{
mlm_assumptions(model)
}
\arguments{
\item{model}{A linear mixed-effects model of class lmerMod or lmerModLmerTest}
}
\value{
TO DO
}
\description{
Reports the results from testing all assumptions of a multilevel model and provides suggestions if an assumption is not passed
}
\references{
TO DO
}
