% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{prints}
\alias{prints}
\alias{print.center}
\alias{print.ICCm}
\alias{print.rsqmlm}
\alias{print.varCompare}
\title{S3Methods for Printing}
\usage{
\method{print}{center}(x, ...)

\method{print}{ICCm}(x, ...)

\method{print}{rsqmlm}(x, ...)

\method{print}{varCompare}(x, ...)
}
\arguments{
\item{x}{Object from \code{mlmtools} package}

\item{...}{Additional arguments}
}
\value{
Prints \code{mlmtools} object
}
\description{
Prints for \code{mlmtools} objects
}
