% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instruction.R
\docType{data}
\name{instruction}
\alias{instruction}
\title{Instruction Data}
\format{
A data frame with 1190 observations on the following 8 variables.
\describe{
\item{\code{female}}{Dummy variable for being female}
\item{\code{mathkind}}{Math achievement score in the spring of kindergarten}
\item{\code{mathgain}}{Gain in math achievement score from spring of kindergarten to spring of first grade}
\item{\code{ses}}{Socioeconomic status}
\item{\code{mathprep}}{First grade teacher's mathematic preparation (based on number of courses taken)}
\item{\code{classid}}{Classroom identifier}
\item{\code{schoolid}}{School identifier}
\item{\code{childid}}{Student identifier}
}
}
\source{
\href{https://www.stata-press.com/data/mlmus3/instruction.dta}{Stata Press}
}
\usage{
data(instruction)
}
\description{
Data from a study on instructional improvement across 312 classrooms.
}
\examples{
data(instruction)
}
\references{
Rabe-Hesketh, Sophia, and Brian Everitt. Handbook of statistical analyses using Stata. CRC Press, 2003.

Hill, H. C., Rowan, B., & Ball, D. L. (2005). Effects of teachers’ mathematical knowledge for teaching on student achievement. American educational research journal, 42(2), 371-406.
}
\keyword{datasets}
