% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelCompare.R
\name{levelCompare}
\alias{levelCompare}
\title{Reports on a likelihood ratio test of the mixed model vs. a model that does not account for nesting}
\usage{
levelCompare(model)
}
\arguments{
\item{model}{A linear mixed-effects model of class lmerMod or lmerModLmerTest}
}
\value{
NEED
}
\description{
Formal test of whether the model that accounts for the correlation of responses within the same unit fits the data better than a model that assumes 0 correlation between responses within the same unit.
}
\details{
NEED
}
\references{
NEED
}
