% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelComparePlot.R
\name{levelComparePlot}
\alias{levelComparePlot}
\title{Plots comparison of accounting for nesting vs. not accounting for nesting}
\usage{
levelComparePlot(
  model,
  x,
  y,
  grouping,
  dataset,
  paneled = TRUE,
  select = c("select"),
  center = FALSE,
  xlab = x,
  ylab = y,
  glab = grouping,
  plot_titles = c("Scatter Plot", "Scatter Plot by Group")
)
}
\arguments{
\item{model}{A linear mixed-effects model of class lmerMod or lmerModLmerTest.}

\item{x}{Predictor variable.}

\item{y}{Outcome variable.}

\item{grouping}{Grouping variable.}

\item{dataset}{A dataset containing the predictor, outcome, and grouping variables.}

\item{paneled}{A logical value indicating whether the plot accounting for nesting should be split into panels.}

\item{select}{A vector indicating the index of the groups to be included in the plots.}

\item{center}{A logical value indicating whether the x variable should be centered}

\item{xlab}{Character vector specifying the horizontal axis label.}

\item{ylab}{Character vector specifying the vertical axis label.}

\item{glab}{Character vector specifying the legend title for the plot accounting for nesting.}

\item{plot_titles}{Character vectors specifying the titles for the plots.}
}
\description{
Plots the line of best for the relationship between two variables accounting for nesting and not accounting for nesting.
}
\examples{
# Gaussian
## Read in data
data(instruction)
## Create model
mod <- lme4::lmer(mathgain ~ mathkind + (1 | classid), data = instruction)
## Generate plots
levelComparePlot(mod, x = "mathkind", y = "mathgain", grouping = "classid", dataset = instruction)

# Logistic
## Read in data
data(reporting)
reporting$final.sample.size <- scale(as.numeric(reporting$final.sample.size))
reporting$mention.outliers <- ifelse(reporting$mention.outliers=="No",0,1)
mod <- lme4::glmer(mention.outliers ~ final.sample.size +
(1 | Journal), data = reporting, family = "binomial")
levelComparePlot(mod, x = "final.sample.size", y = "mention.outliers",
grouping = "Journal", dataset = reporting, paneled = FALSE)


}
